Keo toan bo cong viec cua main (API fixes that: upload minh chung 400, parse
so dang string, ASSIGNED status, finance goal boc lop {goal,progress}; UI kit
dung chung app_input/validators/money_input/empty_state/AppTheme; deep-link
Stripe payment_result + pendingDeepLink giu qua cold-start; Windows runner).

Verify-email: theo dung huong MANDATORY cua main (router ep sang /verify-email
khi chua verify, man verify khoa cung khong nut Để sau, khong tile Profile) —
KHONG giu huong optional cua commit 2d747da. Nhung GIU lai 2 fix trung tinh ve
huong (main thieu ca hai, thieu chung thi chinh luong mandatory cua main cung
hong):
- 403-message: AuthProvider.createFamily tin thang statusCode==403 thay vi
  contains('verif') — vi message that la tieng Viet, contains khong bao gio
  khop -> pendingEmailVerification khong set -> router khong redirect -> luong
  bat buoc verify hong (verify bang kich ban that)
- race-condition: dialog "Da gui yeu cau" refetch family context truoc khi
  dieu huong (ca family_setup + join_family), tranh da member da duoc duyet
  ve /family-setup thay vi dashboard

Verify: flutter analyze 0 error (7 info pre-existing), flutter test 51/51 pass.

+++ lib/widgets/ring_chart.dart
@@ -32,7 +32,7 @@ class RingChart extends StatelessWidget {
           CustomPaint(
             size: Size(size, size),
             painter: _RingPainter(
-              progress: progress.clamp(0.0, 1.0),
+              progress: progress.clamp(0.0, 1.0).toDouble(),
               color: color,
               trackColor: trackColor,
               strokeWidth: strokeWidth,
